+++ src/screen/dashboard.rs
@@ -3363,7 +3363,11 @@ impl Dashboard {
 
         let event = self.file_transfers.receive(request.clone(), config)?;
 
-        self.notifications.notify(
+        let request_attention_window = self
+            .find_window_with_file_transfers()
+            .unwrap_or(self.main_window());
+
+        let request_attention = self.notifications.notify(
             &config.notifications,
             &Notification::FileTransferRequest {
                 nick: request.from.nickname().to_owned(),
@@ -3376,16 +3380,23 @@ impl Dashboard {
                 },
             },
             server,
+            request_attention_window,
         );
 
         let query = target::Query::from(request.from);
 
-        Some(self.handle_file_transfer_event(
+        let task = self.handle_file_transfer_event(
             server,
             &query,
             event,
             &config.buffer,
-        ))
+        );
+
+        if let Some(request_attention) = request_attention {
+            Some(Task::batch(vec![task, request_attention]))
+        } else {
+            Some(task)
+        }
     }
 
     pub fn handle_file_transfer_event(
@@ -3705,13 +3716,37 @@ impl Dashboard {
         }
     }
 
-    pub fn is_open_in_pane(&mut self, kind: &history::Kind) -> bool {
-        self.panes.iter().any(|(_, _, state)| {
+    pub fn find_window_with_file_transfers(&mut self) -> Option<window::Id> {
+        self.panes.iter().find_map(|(window_id, _, state)| {
+            matches!(state.buffer, Buffer::FileTransfers(_))
+                .then_some(window_id)
+        })
+    }
+
+    pub fn find_window_with_history(
+        &mut self,
+        kind: &history::Kind,
+    ) -> Option<window::Id> {
+        self.panes.iter().find_map(|(window_id, _, state)| {
             state
                 .buffer
                 .data()
                 .and_then(history::Kind::from_buffer)
                 .is_some_and(|pane_kind| pane_kind == *kind)
+                .then_some(window_id)
+        })
+    }
+
+    pub fn find_window_with_server(
+        &mut self,
+        server: &Server,
+    ) -> Option<window::Id> {
+        self.panes.iter().find_map(|(window_id, _, state)| {
+            state
+                .buffer
+                .server()
+                .is_some_and(|pane_server| pane_server == *server)
+                .then_some(window_id)
         })
     }